+++ Samples/iOS-Swift/iOS-Swift/Extensions/SpanExtension.swift
@@ -22,8 +22,4 @@ extension Span {
                 
         return self.perform(sel)?.takeUnretainedValue() as? Span
     }
-    
-    func performOnFinish(_ callback : () -> Void) {
-        let object: NSObject
-    }
 }